Move value type utility functions into own file to resolve dependency cycles
