+++ libs/language-server/src/lib/ast/expressions/type-inference.ts
@@ -2,8 +2,6 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-import { strict as assert } from 'assert';
-
 import { assertUnreachable } from 'langium';
 
 import { type ValidationContext } from '../../validation/validation-context';
@@ -36,16 +34,16 @@ import {
   isValuetypeAssignmentLiteral,
 } from '../generated/ast';
 import { getNextAstNodeContainer } from '../model-util';
-// eslint-disable-next-line import/no-cycle
 import {
-  type AtomicValueType,
-  type PrimitiveValueType,
   type ValueType,
   type ValueTypeProvider,
   type WrapperFactoryProvider,
-  isAtomicValueType,
-  isPrimitiveValueType,
 } from '../wrappers';
+import {
+  getValuetypeHierarchyStack,
+  pickCommonAtomicValueType,
+  pickCommonPrimitiveValuetype,
+} from '../wrappers/value-type/util';
 
 import { isEveryValueDefined } from './typeguards';
 
@@ -287,85 +285,6 @@ function inferCollectionElementTypes(
     return undefined;
   }
   return elementValuetypes;
-}
-
-type ValuetypeHierarchyStack = [PrimitiveValueType, ...AtomicValueType[]];
-
-function getValuetypeHierarchyStack(
-  valueType: ValueType,
-): ValuetypeHierarchyStack {
-  if (isPrimitiveValueType(valueType)) {
-    return [valueType];
-  } else if (isAtomicValueType(valueType)) {
-    const supertype = valueType.getSupertype();
-    assert(supertype !== undefined);
-    return [...getValuetypeHierarchyStack(supertype), valueType];
-  }
-  throw new Error(
-    'Should be unreachable, encountered an unknown kind of value type',
-  );
-}
-
-function pickCommonPrimitiveValuetype(
-  primitiveValuetypes: PrimitiveValueType[],
-): PrimitiveValueType | undefined {
-  assert(primitiveValuetypes.length > 0);
-  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-  let resultingType: PrimitiveValueType = primitiveValuetypes[0]!;
-  for (let i = 1; i < primitiveValuetypes.length; ++i) {
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    const currentType = primitiveValuetypes[i]!;
-
-    if (currentType.isConvertibleTo(resultingType)) {
-      continue;
-    }
-
-    if (resultingType.isConvertibleTo(currentType)) {
-      // Pick the more general type as a result
-      resultingType = currentType;
-      continue;
-    }
-
-    // Unable to convert the value types into each other, so there is no common primitive value type
-    return undefined;
-  }
-  return resultingType;
-}
-
-function pickCommonAtomicValueType(
-  stacks: ValuetypeHierarchyStack[],
-): PrimitiveValueType | AtomicValueType | undefined {
-  const minimumStackLength = Math.min(...stacks.map((stack) => stack.length));
-
-  let resultingType: PrimitiveValueType | AtomicValueType | undefined =
-    undefined;
-  for (let stackLevel = 1; stackLevel < minimumStackLength; ++stackLevel) {
-    const typesOfCurrentLevel: (PrimitiveValueType | AtomicValueType)[] =
-      stacks.map(
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        (stack) => stack[stackLevel]!,
-      );
-
-    if (!areAllTypesEqual(typesOfCurrentLevel)) {
-      // Return the common value type of the previous level
-      return resultingType;
-    }
-
-    // Pick any type of the current level since they are all equal
-    resultingType = typesOfCurrentLevel[0];
-  }
-  return resultingType;
-}
-
-function areAllTypesEqual(types: ValueType[]): boolean {
-  for (let i = 1; i < types.length; i++) {
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    if (!types[i - 1]!.equals(types[i]!)) {
-      return false;
-    }
-  }
-
-  return true;
 }
 
 function inferTypeFromValueKeyword(
