+++ libs/language-server/src/lib/ast/wrappers/value-type/index.ts
@@ -2,6 +2,8 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
+export * from './util';
+
 /*
  * Note: Only export types if possible to enforce usage of WrapperFactory outside this directory.
  * This allows us to avoid dependency cycles between the language server and interpreter.
